@@ -1,20 +1,24 @@
 apply plugin: 'com.android.library'
 
 android {
-    compileSdkVersion 27
-    buildToolsVersion "27.0.1"
+    compileSdkVersion 32
     resourcePrefix 'customactivityoncrash_'
 
     defaultConfig {
-        minSdkVersion 14
-        targetSdkVersion 27
-        versionCode 11
-        versionName "2.2.1-SNAPSHOT"
+        minSdkVersion 16
+        targetSdkVersion 32
     }
+    namespace 'cat.ereza.customactivityoncrash'
 }
 
-apply from: 'https://raw.github.com/chrisbanes/gradle-mvn-push/master/gradle-mvn-push.gradle'
+ext {
+    PUBLISH_GROUP_ID = 'cat.ereza'
+    PUBLISH_ARTIFACT_ID = 'customactivityoncrash'
+    PUBLISH_VERSION = '2.4.0-SNAPSHOT'
+}
+
+apply from: "publish-mavencentral.gradle"
 
 dependencies {
-    compile 'com.android.support:appcompat-v7:27.0.1'
+    api 'androidx.appcompat:appcompat:1.4.1'
 }
