@@ -1,15 +1,13 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 27
-    buildToolsVersion "27.0.1"
+    compileSdkVersion 32
 
     defaultConfig {
-        applicationId "cat.ereza.customactivityoncrash.sample"
-        minSdkVersion 14
-        targetSdkVersion 27
-        versionCode 11
-        versionName "2.2.1-SNAPSHOT"
+        minSdkVersion 16
+        targetSdkVersion 32
+        versionCode 13
+        versionName "2.4.0-SNAPSHOT"
     }
     buildTypes {
         debug {
@@ -21,8 +19,9 @@ android {
             proguardFiles 'proguard-rules.pro'
         }
     }
+    namespace 'cat.ereza.customactivityoncrash.sample'
 }
 
 dependencies {
-    compile project(":library")
+    implementation project(":library")
 }
