@@ -16,13 +16,12 @@
 
 package cat.ereza.customactivityoncrash.sample.activity;
 
-import android.annotation.SuppressLint;
 import android.os.AsyncTask;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v7.app.AppCompatActivity;
-import android.view.View;
 import android.widget.Button;
+
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
 
 import cat.ereza.customactivityoncrash.sample.R;
 
@@ -37,42 +36,31 @@ public class MainActivity extends AppCompatActivity {
         Button crashBgThreadButton = findViewById(R.id.button_crash_bg_thread);
         Button crashWithDelayButton = findViewById(R.id.button_crash_with_delay);
 
-        crashMainThreadButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                throw new RuntimeException("I'm a cool exception and I crashed the main thread!");
-            }
+        crashMainThreadButton.setOnClickListener(view -> {
+            throw new RuntimeException("I'm a cool exception and I crashed the main thread!");
         });
 
-        crashBgThreadButton.setOnClickListener(new View.OnClickListener() {
-            @SuppressLint("StaticFieldLeak") //For demo purposes we don't care about leaks
+        //For demo purposes we don't care about leaks nor deprecation
+        //noinspection StaticFieldLeak,deprecation
+        crashBgThreadButton.setOnClickListener(view -> new AsyncTask<Void, Void, Void>() {
             @Override
-            public void onClick(View view) {
-                new AsyncTask<Void, Void, Void>() {
-                    @Override
-                    protected Void doInBackground(Void... voids) {
-                        throw new RuntimeException("I'm also cool, and I crashed the background thread!");
-                    }
-                }.execute();
+            protected Void doInBackground(Void... voids) {
+                throw new RuntimeException("I'm also cool, and I crashed the background thread!");
             }
-        });
+        }.execute());
 
-        crashWithDelayButton.setOnClickListener(new View.OnClickListener() {
-            @SuppressLint("StaticFieldLeak") //For demo purposes we don't care about leaks
+        //For demo purposes we don't care about leaks nor deprecation
+        //noinspection StaticFieldLeak,deprecation
+        crashWithDelayButton.setOnClickListener(view -> new AsyncTask<Void, Void, Void>() {
             @Override
-            public void onClick(View view) {
-                new AsyncTask<Void, Void, Void>() {
-                    @Override
-                    protected Void doInBackground(Void... voids) {
-                        try {
-                            Thread.sleep(5000);
-                        } catch (InterruptedException e) {
-                            //meh
-                        }
-                        throw new RuntimeException("I am a not so cool exception, and I am delayed, so you can check if the app crashes when in background!)");
-                    }
-                }.execute();
+            protected Void doInBackground(Void... voids) {
+                try {
+                    Thread.sleep(5000);
+                } catch (InterruptedException e) {
+                    //meh
+                }
+                throw new RuntimeException("I am a not so cool exception, and I am delayed, so you can check if the app crashes when in background!");
             }
-        });
+        }.execute());
     }
 }
